requirements.txt
langchain
streamlit
python-dotenv
huggingface-hub
langchain-community
langchain-huggingface
cassio
pypdf
astrapy